Modulo (Estoque)

+++ model/veiculo.go
@@ -3,19 +3,19 @@ package model
 import "github.com/go-playground/validator/v10"
 
 type Veiculo struct {
-	Id          int               `json:"id"`
+	Id          int64               `json:"id"`
 	Placa       string            `json:"placa"  validate:"required"`
 	Descricao   string            `json:"descricao"  validate:"required"`
-	CategoriaId int               `json:"categoria_id"  validate:"required"`
+	CategoriaId int64               `json:"categoria_id"  validate:"required"`
 	IsServico   bool              `json:"is_servico"`
 	CreatedAt   DateFormattedTime `json:"created_at"`
 	UpdatedAt   DateFormattedTime `json:"updated_at"`
 }
 
 type VeiculoCategoria struct {
-	Id                    int    `json:"id"`
+	Id                    int64    `json:"id"`
 	Placa                 string `json:"placa"`
-	CategoriaId           int    `json:"categoria_id"`
+	CategoriaId           int64    `json:"categoria_id"`
 	NomeCategoria         string `json:"nome_categoria"`
 	Descricao             string `json:"descricao"` 
 	IsServico             bool   `json:"is_servico"`
@@ -27,7 +27,7 @@ type VeiculoCategoria struct {
 type VeiculoUpdate struct {
 	Placa       string `json:"placa"`
 	Descricao   string `json:"descricao"`
-	CategoriaId int    `json:"categoria_id"`
+	CategoriaId int64    `json:"categoria_id"`
 	IsServico   bool   `json:"is_servico"`
 }
